camera-movement-detection/requirements.txt
pip install streamlit opencv-python numpy Pillow
streamlit
opencv-python
numpy
Pillow
